refactor: LottoMachine 클래스를 service 패키지로 분리
