+++ src/main/java/lotto/Application.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.service.LottoMachine;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
