+++ src/main/java/lotto/service/LottoMachine.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.service;
 
 import java.util.ArrayList;
 import java.util.List;
